Move checkLocation()
